realtime-qa-generator/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests
pydantic
pydantic-settings

firebase-admin
google-cloud-firestore